build: handle auth token at report page
