+++ app/routes.js
@@ -4,4 +4,5 @@ module.exports = routes()
 .add('/', 'dashboard')
 .add('/signin', 'auth')
 .add('/process_auth', 'process_auth')
+.add('/report', 'report')
